refactored/renamed dynamic phrases directories
