+++ app/src/main/kotlin/com/wanderfar/expander/DynamicValue/DynamicValueExampleAdapter.kt
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.wanderfar.expander.DynamicPhrases
+package com.wanderfar.expander.DynamicValue
 
 import android.content.Context
 import android.support.v7.widget.RecyclerView
@@ -24,6 +24,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import com.wanderfar.expander.DynamicPhraseGenerator.DynamicPhrase
+import com.wanderfar.expander.DynamicPhraseGenerator.DynamicPhraseGenerator
 import com.wanderfar.expander.R
 import java.util.*
