+++ app/src/main/kotlin/com/wanderfar/expander/DynamicValue/TextDrawableCustom.kt
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.wanderfar.expander.DynamicPhrases
+package com.wanderfar.expander.DynamicValue
 
 import android.graphics.*
 import android.graphics.drawable.Drawable
