+++ app/src/main/res/layout/activity_macro.xml
@@ -54,7 +54,7 @@
                 android:paddingStart="?android:attr/scrollbarSize"
                 android:paddingEnd="?android:attr/scrollbarSize">
 
-                <com.wanderfar.expander.DynamicPhrases.DynamicValueEditText
+                <com.wanderfar.expander.DynamicValue.DynamicValueEditText
                     xmlns:custom="http://schemas.android.com/apk/res/com.wanderfar.expander"
                     android:id="@+id/input_phrase"
                     android:layout_width="match_parent"
